requirements.txt
annotated-types==0.7.0
blinker==1.9.0
bunnet==1.3.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
Flask==3.1.2
Flask-Mail==0.10.0
Flask-WTF==1.2.2
hiredis==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy-model==0.2.0
MarkupSafe==3.0.3
more-itertools==10.8.0
passlib==1.7.4
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pymongo==4.15.3
python-dotenv==1.2.1
python-ulid==1.1.0
redis==5.3.1
redis-om==0.3.5
setuptools==80.9.0
toml==0.10.2
types-cffi==1.17.0.20250915
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==80.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
